refactor(infra-billing): dedupe the site-ownership check shared by GET and PORTAL
INFRA_BILLING_GET and INFRA_BILLING_PORTAL each hand-rolled the identical
lowercase+dedupe-slugs, load-owned-slugs, reject-unowned check. Extract it
into resolveOwnedSlugs() in a new ownership.ts and call it from both.

Net: -20 lines duplicated logic / +23 lines shared helper, behavior
unchanged (byte-identical logic, just relocated + parameterized on ctx/orgId).
